removing tf-cnn and fixing the poor scores of torch cnn
